fix: android layout issues
