+++ .github/workflows/symanticrelease.yml
@@ -11,7 +11,7 @@ env:
   DOTNET_NOLOGO: true                     # Disable the .NET logo
   DOTNET_SKIP_FIRST_TIME_EXPERIENCE: true # Disable the .NET first time experience
   DOTNET_CLI_TELEMETRY_OPTOUT: true       # Disable sending .NET CLI telemetry
-  DONET_VERSION: 9.0.100
+  DONET_VERSION: 9.0.200
  
 jobs:
   build:
@@ -25,7 +25,7 @@ jobs:
      - name: Install MAUI workload
        run: dotnet workload install maui
      
-     - uses: actions/checkout@v3
+     - uses: actions/checkout@v4
      - name: Build
        run: dotnet build src/Indiko.Maui.Controls.Chat.sln -c Release
