+++ samples/Indiko.Maui.Controls.Chat.Sample.sln
@@ -9,7 +9,6 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9C4DBE8F-19C9-468A-A26A-A4B975098C6F}"
 	ProjectSection(SolutionItems) = preProject
 		Brainstorm.md = Brainstorm.md
-		CurrentFeatures.md = CurrentFeatures.md
 		..\README.md = ..\README.md
 		..\.github\workflows\release-nuget.yml = ..\.github\workflows\release-nuget.yml
 		..\.github\workflows\symanticrelease.yml = ..\.github\workflows\symanticrelease.yml
